FraudDetect/proctorAI/requirements.txt
# Computer Vision
opencv-python==4.9.0.80
mediapipe==0.10.9
face-recognition==1.3.0
#dlib-bin>=20.0.1
dlib==19.24.1
#https://github.com/z-mahmud22/Dlib_Windows_Python3.x/raw/main/dlib-19.24.1-cp311-cp311-win_amd64.whlwebrtcvad-wheels
# ML / Deep Learning
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
scikit-learn==1.4.2

# Audio
pyaudio==0.2.14
librosa==0.10.2
#webrtcvad==2.0.10
webrtcvad-wheels

# Behavioral Biometrics
pynput==1.7.6

# WebSocket (to send events to backend)
websockets==12.0

# Utilities
python-dotenv==1.0.1
pillow==10.3.0
scipy==1.13.0

pyobjc-framework-Cocoa; sys_platform == "darwin"
pywin32; sys_platform == "win32"
psutil